requirements.txt
attrs==21.2.0
autopep8==1.5.7
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography==3.4.7
cycler==0.10.0
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
dill==0.3.4
dlx==1.0.4
docplex==2.20.204
fastdtw==0.3.4
fastjsonschema==2.15.1
h5py==3.3.0
idna==2.10
inflection==0.5.1
ipykernel @ file:///C:/ci/ipykernel_1607454116140/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1617120911285/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///C:/ci/jedi_1606932640618/work
joblib==1.0.1
jsonschema==3.2.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213515523/work
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
more-itertools==8.8.0
mpmath==1.2.1
multitasking==0.0.9
ntlm-auth==1.5.0
numpy==1.21.0
pandas==1.2.5
parso==0.7.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.2.0
ply==3.11
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
psutil==5.8.0
pybind11==2.6.2
pycodestyle==2.7.0
pycparser==2.20
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
pyparsing==2.4.7
pyrsistent==0.17.3
python-constraint==1.4.0
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-resize-image==1.1.19
pytz==2021.1
pywin32==228
pyzmq==20.0.0
qiskit==0.27.0
qiskit-aer==0.8.2
qiskit-aqua==0.9.2
qiskit-ibmq-provider==0.14.0
qiskit-ignis==0.6.0
qiskit-terra==0.17.4
Quandl==3.6.1
requests==2.25.1
requests-ntlm==1.1.0
retworkx==0.9.0
scikit-learn==0.24.2
scipy==1.7.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sympy==1.8
threadpoolctl==2.1.0
toml==0.10.2
tornado @ file:///C:/ci/tornado_1606924294691/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
urllib3==1.26.5
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
websocket-client==1.1.0
wincertstore==0.2
yfinance==0.1.59
